small postgres fixes

+++ src/main/java/uk/ac/ed/inf/acp/model/Capability.java
@@ -3,16 +3,16 @@ package uk.ac.ed.inf.acp.model;
 public class Capability {
     Boolean cooling;
     Boolean heating;
-    Integer capacity;
-    Integer maxMoves;
-    Integer costPerMove;
-    Integer costInitial;
-    Integer costFinal;
+    Double capacity;
+    Double maxMoves;
+    Double costPerMove;
+    Double costInitial;
+    Double costFinal;
 
     Capability(){
 
     }
-    Capability (Boolean cooling, Boolean heating, Integer capacity, Integer maxMoves, Integer costPerMove, Integer costInitial, Integer costFinal) {
+    Capability (Boolean cooling, Boolean heating, Double capacity, Double maxMoves, Double costPerMove, Double costInitial, Double costFinal) {
         this.cooling = cooling;
         this.heating = heating;
         this.capacity = capacity != null ? capacity : 0;
@@ -35,34 +35,34 @@ public class Capability {
     public void setHeating(Boolean heating) {
         this.heating = heating;
     }
-    public Integer getCapacity() {
+    public Double getCapacity() {
         return capacity;
     }
-    public void setCapacity(Integer capacity) {
+    public void setCapacity(Double capacity) {
         this.capacity = capacity;
     }
-    public Integer getMaxMoves() {
+    public Double getMaxMoves() {
         return maxMoves;
     }
-    public void setMaxMoves(Integer maxMoves) {
+    public void setMaxMoves(Double maxMoves) {
         this.maxMoves = maxMoves;
     }
-    public Integer getCostPerMove() {
+    public Double getCostPerMove() {
         return costPerMove;
     }
-    public void setCostPerMove(Integer costPerMove) {
+    public void setCostPerMove(Double costPerMove) {
         this.costPerMove = costPerMove;
     }
-    public Integer getCostInitial() {
+    public Double getCostInitial() {
         return costInitial;
     }
-    public void setCostInitial(Integer costInitial) {
+    public void setCostInitial(Double costInitial) {
         this.costInitial = costInitial;
     }
-    public Integer getCostFinal() {
+    public Double getCostFinal() {
         return costFinal;
     }
-    public void setCostFinal(Integer costFinal) {
+    public void setCostFinal(Double costFinal) {
         this.costFinal = costFinal;
     }
 }
